+++ test/Psr7ServerRequestTest.php
@@ -1,22 +1,21 @@
 <?php
+
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @see       http://github.com/zendframework/zend-diactoros for the canonical source repository
- * @copyright Copyright (c) 2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   https://github.com/zendframework/zend-diactoros/blob/master/LICENSE.md New BSD License
+ * @see       https://github.com/laminas/laminas-psr7bridge for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-psr7bridge/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-psr7bridge/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Psr7Bridge;
+namespace LaminasTest\Psr7Bridge;
 
+use Laminas\Diactoros\ServerRequest;
+use Laminas\Diactoros\UploadedFile;
+use Laminas\Psr7Bridge\Psr7ServerRequest;
 use PHPUnit_Framework_TestCase as TestCase;
-use Zend\Diactoros\ServerRequest;
-use Zend\Diactoros\UploadedFile;
-use Zend\Psr7Bridge\Psr7ServerRequest;
 
 class Psr7ServerRequestTest extends TestCase
 {
-    public function testToZendWithShallowOmitsBody()
+    public function testToLaminasWithShallowOmitsBody()
     {
         $server = [
             'SCRIPT_NAME'     => __FILE__,
@@ -64,47 +63,47 @@ class Psr7ServerRequestTest extends TestCase
             ->withCookieParams($cookies)
             ->withParsedBody($bodyParams);
 
-        $zendRequest = Psr7ServerRequest::toZend($psr7Request, $shallow = true);
+        $laminasRequest = Psr7ServerRequest::toLaminas($psr7Request, $shallow = true);
 
-        // This needs to be a ZF2 request
-        $this->assertInstanceOf('Zend\Http\PhpEnvironment\Request', $zendRequest);
-        $this->assertInstanceOf('Zend\Http\Request', $zendRequest);
+        // This needs to be a Laminas request
+        $this->assertInstanceOf('Laminas\Http\PhpEnvironment\Request', $laminasRequest);
+        $this->assertInstanceOf('Laminas\Http\Request', $laminasRequest);
 
         // But, more specifically, an instance where we do not use superglobals
         // to inject it
-        $this->assertInstanceOf('Zend\Psr7Bridge\Zend\Request', $zendRequest);
+        $this->assertInstanceOf('Laminas\Psr7Bridge\Laminas\Request', $laminasRequest);
 
         // Assert shallow conditions
         // (content, files, and body parameters are not injected)
-        $this->assertEmpty($zendRequest->getContent());
-        $this->assertCount(0, $zendRequest->getFiles());
-        $this->assertCount(0, $zendRequest->getPost());
+        $this->assertEmpty($laminasRequest->getContent());
+        $this->assertCount(0, $laminasRequest->getFiles());
+        $this->assertCount(0, $laminasRequest->getPost());
 
         // Assert all other Request metadata
-        $this->assertEquals($uri, $zendRequest->getRequestUri());
-        $this->assertEquals($method, $zendRequest->getMethod());
+        $this->assertEquals($uri, $laminasRequest->getRequestUri());
+        $this->assertEquals($method, $laminasRequest->getMethod());
 
-        $zf2Headers = $zendRequest->getHeaders();
-        $this->assertTrue($zf2Headers->has('Host'));
-        $this->assertTrue($zf2Headers->has('X-Foo'));
-        $this->assertTrue($zf2Headers->has('Content-Type'));
-        $this->assertEquals('example.com', $zf2Headers->get('Host')->getFieldValue());
-        $this->assertEquals('bar', $zf2Headers->get('X-Foo')->getFieldValue());
-        $this->assertEquals('multipart/form-data', $zf2Headers->get('Content-Type')->getFieldValue());
+        $laminasHeaders = $laminasRequest->getHeaders();
+        $this->assertTrue($laminasHeaders->has('Host'));
+        $this->assertTrue($laminasHeaders->has('X-Foo'));
+        $this->assertTrue($laminasHeaders->has('Content-Type'));
+        $this->assertEquals('example.com', $laminasHeaders->get('Host')->getFieldValue());
+        $this->assertEquals('bar', $laminasHeaders->get('X-Foo')->getFieldValue());
+        $this->assertEquals('multipart/form-data', $laminasHeaders->get('Content-Type')->getFieldValue());
 
-        $this->assertTrue($zf2Headers->has('Cookie'));
-        $cookie = $zf2Headers->get('Cookie');
-        $this->assertInstanceOf('Zend\Http\Header\Cookie', $cookie);
+        $this->assertTrue($laminasHeaders->has('Cookie'));
+        $cookie = $laminasHeaders->get('Cookie');
+        $this->assertInstanceOf('Laminas\Http\Header\Cookie', $cookie);
         $this->assertTrue(isset($cookie['PHPSESSID']));
         $this->assertEquals($cookies['PHPSESSID'], $cookie['PHPSESSID']);
 
-        $test = $zendRequest->getServer();
+        $test = $laminasRequest->getServer();
         $this->assertCount(2, $test);
         $this->assertEquals(__FILE__, $test->get('SCRIPT_NAME'));
         $this->assertEquals(__FILE__, $test->get('SCRIPT_FILENAME'));
     }
 
-    public function testCanCastFullRequestToZend()
+    public function testCanCastFullRequestToLaminas()
     {
         $server = [
             'SCRIPT_NAME'     => __FILE__,
@@ -152,36 +151,36 @@ class Psr7ServerRequestTest extends TestCase
             ->withCookieParams($cookies)
             ->withParsedBody($bodyParams);
 
-        $zendRequest = Psr7ServerRequest::toZend($psr7Request);
+        $laminasRequest = Psr7ServerRequest::toLaminas($psr7Request);
 
-        // This needs to be a ZF2 request
-        $this->assertInstanceOf('Zend\Http\PhpEnvironment\Request', $zendRequest);
-        $this->assertInstanceOf('Zend\Http\Request', $zendRequest);
+        // This needs to be a Laminas request
+        $this->assertInstanceOf('Laminas\Http\PhpEnvironment\Request', $laminasRequest);
+        $this->assertInstanceOf('Laminas\Http\Request', $laminasRequest);
 
         // But, more specifically, an instance where we do not use superglobals
         // to inject it
-        $this->assertInstanceOf('Zend\Psr7Bridge\Zend\Request', $zendRequest);
+        $this->assertInstanceOf('Laminas\Psr7Bridge\Laminas\Request', $laminasRequest);
 
-        $this->assertEquals($uri, $zendRequest->getRequestUri());
-        $this->assertEquals($method, $zendRequest->getMethod());
+        $this->assertEquals($uri, $laminasRequest->getRequestUri());
+        $this->assertEquals($method, $laminasRequest->getMethod());
 
-        $zf2Headers = $zendRequest->getHeaders();
-        $this->assertTrue($zf2Headers->has('Host'));
-        $this->assertTrue($zf2Headers->has('X-Foo'));
-        $this->assertTrue($zf2Headers->has('Content-Type'));
-        $this->assertEquals('example.com', $zf2Headers->get('Host')->getFieldValue());
-        $this->assertEquals('bar', $zf2Headers->get('X-Foo')->getFieldValue());
-        $this->assertEquals('multipart/form-data', $zf2Headers->get('Content-Type')->getFieldValue());
+        $laminasHeaders = $laminasRequest->getHeaders();
+        $this->assertTrue($laminasHeaders->has('Host'));
+        $this->assertTrue($laminasHeaders->has('X-Foo'));
+        $this->assertTrue($laminasHeaders->has('Content-Type'));
+        $this->assertEquals('example.com', $laminasHeaders->get('Host')->getFieldValue());
+        $this->assertEquals('bar', $laminasHeaders->get('X-Foo')->getFieldValue());
+        $this->assertEquals('multipart/form-data', $laminasHeaders->get('Content-Type')->getFieldValue());
 
-        $this->assertTrue($zf2Headers->has('Cookie'));
-        $cookie = $zf2Headers->get('Cookie');
-        $this->assertInstanceOf('Zend\Http\Header\Cookie', $cookie);
+        $this->assertTrue($laminasHeaders->has('Cookie'));
+        $cookie = $laminasHeaders->get('Cookie');
+        $this->assertInstanceOf('Laminas\Http\Header\Cookie', $cookie);
         $this->assertTrue(isset($cookie['PHPSESSID']));
         $this->assertEquals($cookies['PHPSESSID'], $cookie['PHPSESSID']);
 
-        $this->assertEquals(file_get_contents(__FILE__), (string) $zendRequest->getContent());
+        $this->assertEquals(file_get_contents(__FILE__), (string) $laminasRequest->getContent());
 
-        $test = $zendRequest->getFiles();
+        $test = $laminasRequest->getFiles();
         $this->assertCount(1, $test);
         $this->assertTrue(isset($test['foo']));
         $upload = $test->get('foo');
@@ -191,20 +190,20 @@ class Psr7ServerRequestTest extends TestCase
         $this->assertArrayHasKey('tmp_name', $upload);
         $this->assertArrayHasKey('error', $upload);
 
-        $this->assertEquals($bodyParams, $zendRequest->getPost()->getArrayCopy());
+        $this->assertEquals($bodyParams, $laminasRequest->getPost()->getArrayCopy());
 
-        $test = $zendRequest->getServer();
+        $test = $laminasRequest->getServer();
         $this->assertCount(2, $test);
         $this->assertEquals(__FILE__, $test->get('SCRIPT_NAME'));
         $this->assertEquals(__FILE__, $test->get('SCRIPT_FILENAME'));
     }
 
-    public function testNestedFileParametersArePassedCorrectlyToZendRequest()
+    public function testNestedFileParametersArePassedCorrectlyToLaminasRequest()
     {
         $this->markTestIncomplete('Functionality is written but untested');
     }
 
-    public function testCustomHttpMethodsDoNotRaiseAnExceptionDuringConversionToZendRequest()
+    public function testCustomHttpMethodsDoNotRaiseAnExceptionDuringConversionToLaminasRequest()
     {
         $this->markTestIncomplete('Functionality is written but untested');
     }